Refactor test namespaces and update project file naming for clarity
