+++ Providers/tests/Unit/BingProviderTests.cs
@@ -1,6 +1,6 @@
 // Copyright (c) 2014-2025 Sarin Na Wangkanai, All Rights Reserved. Apache License, Version 2.0
 
-namespace Wangkanai.Planet.Providers.Bing.Tests;
+namespace Wangkanai.Planet.Providers;
 
 public class BingProviderTests
 {
