+++ Providers/tests/Unit/Wangkanai.Planet.Providers.UnitTests.csproj
